Add Model::*NestedNameExists and *ByNestedName
These are extensions of the *NameExists and *ByName APIs
(like LinkNameExists and LinkByName) that allow passing
a nested name that can begin with a sequence of nested model
names separated by :: and may end with the name of an object
of the specified type, such as "outer_model::inner_model::inner_joint".
An example model file and tests are included.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>