surface.sdf: add rolling friction parameters
Rolling friction differs from tangential and torsional
friction in that it is not caused by shear stress
between bodies in contact. Rather it is the result
of a shift in the centroid of normal contact
stresses caused by deformation of the rolling body.

The two parameters added to //surface/friction/rolling are:
* deformation_length: a length representing the displacement
  of the centroid of a contact pressure distribution. This
  parameter the rolling friction parameter in MJCF.
* coefficient: a non-dimensional measure of the deformation
  length scaled by the object size. This is equivalent to
  the Rolling Resistance Coefficient used in automotive
  engineering.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>